requirements.txt
fastapi
uvicorn
tensorflow-intel==2.12.0
numpy==1.23.5
pillow
python-multipart
openai
python-dotenv
